requirements.txt
fastapi
uvicorn
jinja2
python-multipart
groq
pymupdf
pillow
pytesseract
sentence-transformers
torch
pandas